BAEL-3656 Change the integration test into unit test.
